+++ hdt-java-core/src/main/java/org/rdfhdt/hdt/cache/DictionaryCacheLRI.java
@@ -39,22 +39,22 @@ import java.util.concurrent.ConcurrentHashMap;
  */
 public class DictionaryCacheLRI<T> implements DictionaryCache<T> {
 
-	private Map<Integer, T> cache;
-	private int [] arr;
+	private Map<Long, T> cache;
+	private long [] arr;
 	private int ptr=0;
 	private final int size;
 	
 	public DictionaryCacheLRI(int size) {
 		this.size = size;
-		arr = new int[size];
-		cache = new ConcurrentHashMap<Integer, T>(size);
+		arr = new long[size];
+		cache = new ConcurrentHashMap<Long, T>(size);
 	}
 		
 	/* (non-Javadoc)
 	 * @see hdt.jena.DictionaryNodeCache#getNode(int)
 	 */
 	@Override
-	public T get(int id) {
+	public T get(long id) {
 		return cache.get(id);
 	}
 
@@ -62,7 +62,7 @@ public class DictionaryCacheLRI<T> implements DictionaryCache<T> {
 	 * @see hdt.jena.DictionaryNodeCache#setNode(int, com.hp.hpl.jena.graph.Node)
 	 */
 	@Override
-	public void put(int id, T node) {
+	public void put(long id, T node) {
 		cache.put(id, node);
 		if(cache.size()>size) {
 			cache.remove(arr[ptr]);
